[fmt] Fix incorrect format specifiers in printf
* Since tm.* members are ints, apply cast to unsigned int to format
  them correctly via printf.
* I64d(u) is not correct format specifier for long long (unsigned),
  since such types can be more than 64 bit long. Either use platform
  -specific (u)int64_t, or, better, change format specifier to %lld /
  %llu. See https://stackoverflow.com/questions/15718039/why-the-use-of-cin-cout-or-i64d-is-preferred-over-lld-specifier-in-c for details.

+++ src/bled/header_verbose_list.c
@@ -33,13 +33,13 @@ void FAST_FUNC header_verbose_list(const file_header_t *file_header)
 		bb_mode_string(file_header->mode),
 		user,
 		group,
-		file_header->size,
-		1900 + ptm->tm_year,
-		1 + ptm->tm_mon,
-		ptm->tm_mday,
-		ptm->tm_hour,
-		ptm->tm_min,
-		ptm->tm_sec,
+		(unsigned long long)file_header->size,
+		1900U + (unsigned)ptm->tm_year,
+		1U + (unsigned)ptm->tm_mon,
+		(unsigned)ptm->tm_mday,
+		(unsigned)ptm->tm_hour,
+		(unsigned)ptm->tm_min,
+		(unsigned)ptm->tm_sec,
 		file_header->name);
 
 #else /* !FEATURE_TAR_UNAME_GNAME */
@@ -50,13 +50,13 @@ void FAST_FUNC header_verbose_list(const file_header_t *file_header)
 		bb_mode_string(file_header->mode),
 		(unsigned)file_header->uid,
 		(unsigned)file_header->gid,
-		file_header->size,
-		1900 + ptm->tm_year,
-		1 + ptm->tm_mon,
-		ptm->tm_mday,
-		ptm->tm_hour,
-		ptm->tm_min,
-		ptm->tm_sec,
+		(unsigned long long)file_header->size,
+		1900U + (unsigned)ptm->tm_year,
+		1U + (unsigned)ptm->tm_mon,
+		(unsigned)ptm->tm_mday,
+		(unsigned)ptm->tm_hour,
+		(unsigned)ptm->tm_min,
+		(unsigned)ptm->tm_sec,
 		file_header->name);
 
 #endif /* FEATURE_TAR_UNAME_GNAME */
